functions/spotify/requirements.txt
aiohttp==3.8.4 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.6.2 ; python_version >= "3.10" and python_version < "4"
cachetools==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.4 ; python_version >= "3.10" and python_version < "4"
cloudevents==1.9.0 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
db-dtypes==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4"
et-xmlfile==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
flask==2.3.2 ; python_version >= "3.10" and python_version < "4"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2023.6.0 ; python_version >= "3.10" and python_version < "4.0"
functions-framework==3.4.0 ; python_version >= "3.10" and python_version < "4"
gcsfs==2023.6.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth-oauthlib==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.21.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-bigquery-storage==2.22.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-bigquery==3.11.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-functions==1.13.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.59.1 ; python_version < "4.0" and python_version >= "3.10"
googleapis-common-protos[grpc]==1.59.1 ; python_version >= "3.10" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.56.0 ; python_version < "4.0" and python_version >= "3.10"
grpcio==1.56.0 ; python_version < "4.0" and python_version >= "3.10"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.25.1 ; python_version < "4.0" and python_version >= "3.10"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
openpyxl==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pandas-gbq==0.19.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.23.4 ; python_version < "4.0" and python_version >= "3.10"
pyarrow==12.0.1 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
pydata-google-auth==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.10" and python_version < "4.0"
watchdog==3.0.0 ; python_version >= "3.10" and python_version < "4"
werkzeug==2.3.6 ; python_version >= "3.10" and python_version < "4"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "4.0"
